move Product to PIM package
